Rework Cutouts and Measurements (#302)
Changed the data-product structure so that there is one Cutouts per SourceList, which has an attribute `co_dict` that is a dictionary (actually a class that inherits from dictionary) of dictionaries containing data for all cutouts used by measurements. This data is saved and loaded from disk using the filepath attribute which is saved to the database.

Added attributes to Measurements that can access this data (eg. `Measurements.sub_data`) through the Cutouts relationship, loading the relevant data when needed from disk.